+++ R/plot_effects.R
@@ -96,7 +96,7 @@ NULL
 #' @rdname calendar_effect_plots
 #' @export
 plot_day_of_week_effects <- function(x, type = c("epidemic", "report"),
-                                     measure = c("normalized", "percent"), ...) {
+                                     measure = c("percent", "normalized"), ...) {
   .tbl_now_plot_panel(x, .tbl_now_effect_key("weekday", match.arg(type)),
                       measure = match.arg(measure), ...)
 }
@@ -104,7 +104,7 @@ plot_day_of_week_effects <- function(x, type = c("epidemic", "report"),
 #' @rdname calendar_effect_plots
 #' @export
 plot_week_of_year_effects <- function(x, type = c("epidemic", "report"),
-                                      measure = c("normalized", "percent"), ...) {
+                                      measure = c("percent", "normalized"), ...) {
   .tbl_now_plot_panel(x, .tbl_now_effect_key("week", match.arg(type)),
                       measure = match.arg(measure), ...)
 }
@@ -112,7 +112,7 @@ plot_week_of_year_effects <- function(x, type = c("epidemic", "report"),
 #' @rdname calendar_effect_plots
 #' @export
 plot_month_of_year_effects <- function(x, type = c("epidemic", "report"),
-                                       measure = c("normalized", "percent"), ...) {
+                                       measure = c("percent", "normalized"), ...) {
   .tbl_now_plot_panel(x, .tbl_now_effect_key("month", match.arg(type)),
                       measure = match.arg(measure), ...)
 }
@@ -120,7 +120,7 @@ plot_month_of_year_effects <- function(x, type = c("epidemic", "report"),
 #' @rdname calendar_effect_plots
 #' @export
 plot_holiday_effects <- function(x, type = c("epidemic", "report"),
-                                 measure = c("normalized", "percent"), ...) {
+                                 measure = c("percent", "normalized"), ...) {
   .tbl_now_plot_panel(x, .tbl_now_effect_key("holiday", match.arg(type)),
                       measure = match.arg(measure), ...)
 }
@@ -128,7 +128,7 @@ plot_holiday_effects <- function(x, type = c("epidemic", "report"),
 #' @rdname calendar_effect_plots
 #' @export
 plot_holiday_lag_effects <- function(x, type = c("epidemic", "report"),
-                                     measure = c("normalized", "percent"), ...) {
+                                     measure = c("percent", "normalized"), ...) {
   .tbl_now_plot_panel(x, .tbl_now_effect_key("holiday_lag", match.arg(type)),
                       measure = match.arg(measure), ...)
 }